LPD-91681 Add test
