+++ modules/apps/calendar/calendar-test/build.gradle
@@ -12,7 +12,10 @@ dependencies {
 	testIntegrationImplementation project(":apps:portal-search:portal-search-api")
 	testIntegrationImplementation project(":apps:portal-search:portal-search-test-util")
 	testIntegrationImplementation project(":apps:portal:portal-upgrade-test-util")
+	testIntegrationImplementation project(":apps:portlet-display-template:portlet-display-template-api")
 	testIntegrationImplementation project(":apps:ratings:ratings-test-util")
 	testIntegrationImplementation project(":apps:static:portal:portal-upgrade-api")
+	testIntegrationImplementation project(":apps:template:template-api")
+	testIntegrationImplementation project(":apps:template:template-test-util")
 	testIntegrationImplementation project(":test:arquillian-extension-junit-bridge")
 }
